Local Grid/requirements.txt
selenium>=4.9.0
pytest-selenium
pytest-parallel